requirements.txt
mysql-connector-python>=8.0
customtkinter==5.2.2
Pillow>=9.0.0
openpyxl>=3.0.0